requirements.txt
Django==5.1.4
django-debug-toolbar==4.4.6
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
geographiclib==2.0
geopy==2.4.1
pillow==11.0.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10
pytz==2024.2
timezonefinder==6.5.7
tzdata==2024.2